Move form components into dedicated directory
