+++ assets/js/vue-apps/form-components/index.js
@@ -3,8 +3,8 @@ import Vue from 'vue';
 import $ from 'jquery';
 import forEach from 'lodash/forEach';
 
-import WordCount from './components/word-count.vue';
-import BudgetInput from './components/budget-input.vue';
+import WordCount from './word-count.vue';
+import BudgetInput from './budget-input.vue';
 
 function initWordCount() {
     forEach(document.querySelectorAll('.js-word-count'), el => {
